@@ -10,9 +10,9 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
-        "laravel/framework": ">=8.0",
-        "ext-json": "*"
+        "php": "^8.1",
+        "ext-json": "*",
+        "laravel/framework": ">=10.0"
     },
     "autoload": {
         "psr-4": {
@@ -26,6 +26,6 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.0"
+        "phpunit/phpunit": "^10.0"
     }
 }
