@@ -3,6 +3,7 @@
 namespace Codewiser\Workflow\Example;
 
 use Codewiser\Workflow\Charge;
+use Codewiser\Workflow\Context;
 use Codewiser\Workflow\Exceptions\TransitionFatalException;
 use Codewiser\Workflow\Exceptions\TransitionRecoverableException;
 use Codewiser\Workflow\Transition;
@@ -13,6 +14,8 @@ use Codewiser\Workflow\WorkflowBlueprint;
  */
 class ArticleEnumWorkflow extends \Codewiser\Workflow\WorkflowBlueprint
 {
+    protected static int $charge = 0;
+
     public function states(): array
     {
         return Enum::cases();
@@ -22,37 +25,31 @@ class ArticleEnumWorkflow extends \Codewiser\Workflow\WorkflowBlueprint
     {
         return [
             Transition::make(Enum::new, Enum::review)
-                ->before(function (Article $model) {
-                    throw new TransitionRecoverableException();
-                })
+                ->before(fn() => throw new TransitionRecoverableException())
                 ->set('color', 'red'),
 
             Transition::make(Enum::review, Enum::published)->as('Fatal transition')
-                ->before(function (Article $model) {
-                    throw new TransitionFatalException();
-                }),
+                ->before(fn() => throw new TransitionFatalException()),
 
             Transition::make(Enum::review, Enum::correction)
                 ->rules([
                     'comment' => 'required'
                 ])
                 ->authorizedBy([$this, 'authorize'])
-                ->after(function (Article $model, array $context) {
-                    $model->body = $context['comment'];
+                ->after(function (ArticleWithEnum $model, Context $context) {
+                    $model->body = $context->data()->get('comment');
                 }),
 
             Transition::make(Enum::correction, Enum::review)
-                ->authorizedBy(function (Article $model) {
-                    return false;
-                }),
+                ->authorizedBy(fn() => false),
 
             Transition::make(Enum::new, Enum::cumulative)
                 ->chargeable(Charge::make(
-                    function (Article $model) {
-                        return 1 / 3;
+                    function () {
+                        return self::$charge / 3;
                     },
-                    function (Article $model) {
-                        //
+                    function () {
+                        self::$charge++;
                     }
                 ))
 
