@@ -7,30 +7,14 @@ use Illuminate\Support\MultipleItemsFoundException;
 
 class BlueprintValidator
 {
-    /**
-     * @var StateCollection
-     */
-    public $states;
+    public StateCollection $states;
 
-    /**
-     * @var TransitionCollection
-     */
-    public $transitions;
+    public TransitionCollection $transitions;
 
-    /**
-     * @var bool
-     */
-    public $valid = true;
+    public bool $valid = true;
 
-    /**
-     * @var WorkflowBlueprint
-     */
-    public $blueprint;
-
-    public function __construct(WorkflowBlueprint $blueprint)
+    public function __construct(public WorkflowBlueprint $blueprint)
     {
-        $this->blueprint = $blueprint;
-
         $this->states = StateCollection::make($blueprint->states());
 
         $this->transitions = TransitionCollection::make($blueprint->transitions());
@@ -41,8 +25,8 @@ class BlueprintValidator
         return $this->transitions
             ->map(function (Transition $transition) {
                 $row = [
-                    'source' => Value::scalar($transition->source),
-                    'target' => Value::scalar($transition->target),
+                    'source' => $transition->source->name,
+                    'target' => $transition->target->name,
                     'caption' => $transition->caption ?? $this->states->one($transition->target)->caption(),
                     'prerequisites' => $transition->prerequisites()->isEmpty() ? 'No' : 'Yes',
                     'authorization' => is_null($transition->authorization()) ? 'No' : 'Yes',
@@ -76,7 +60,7 @@ class BlueprintValidator
         return $this->states
             ->map(function (State $state) {
                 $row = [
-                    'value' => Value::scalar($state),
+                    'value' => $state->value->name,
                     'caption' => $state->caption(),
                     'additional' => json_encode($state->additional()),
                     'error' => null
