@@ -4,7 +4,6 @@ namespace Codewiser\Workflow\Events;
 
 use Codewiser\Workflow\Context;
 use Codewiser\Workflow\StateMachineEngine;
-use Codewiser\Workflow\Transition;
 use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Foundation\Events\Dispatchable;
@@ -17,25 +16,10 @@ class ModelTransited
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
-    /**
-     * @var StateMachineEngine
-     */
-    public $engine;
+    public Model $model;
 
-    /**
-     * @var Model
-     */
-    public $model;
-
-    /**
-     * @var Context
-     */
-    public $context;
-
-    public function __construct(StateMachineEngine $engine, Context $context)
+    public function __construct(public StateMachineEngine $engine, public Context $context)
     {
-        $this->engine = $engine;
-        $this->model = $engine->model;
-        $this->context = $context;
+        $this->model = $this->engine->model;
     }
 }
