@@ -5,14 +5,11 @@ namespace Codewiser\Workflow\Exceptions;
 use Throwable;
 
 /**
- * Throws this exception to prevent transition from being in list of relevant transitions.
+ * Throws this exception to prevent transition from being in a list of relevant transitions.
  */
 class TransitionFatalException extends TransitionException
 {
-    /**
-     * @var int
-     */
-    public $status = 403;
+    public int $status = 403;
 
     public function __construct($message = "Transition is disabled", $code = 0, Throwable $previous = null)
     {
