@@ -22,25 +22,11 @@ use Illuminate\Validation\ValidationException;
  * Transition between states in State Machine.
  *
  * @template TModel of \Illuminate\Database\Eloquent\Model
- * @template TType
+ * @template TType of \UnitEnum
  */
 class Transition implements Arrayable, Injectable
 {
     use HasAttributes, HasStateMachineEngine, HasCaption, HasCallbacks, HasValidationRules, HasPrerequisites, HasFootprint;
-
-    /**
-     * Source state.
-     *
-     * @var TType
-     */
-    public $source;
-
-    /**
-     * Target state.
-     *
-     * @var TType
-     */
-    public $target;
 
     /**
      * Instructions to authorize transit.
@@ -59,24 +45,22 @@ class Transition implements Arrayable, Injectable
     /**
      * Instantiate new transition.
      *
-     * @param  TType  $source
-     * @param  TType  $target
+     * @param  TType&\UnitEnum  $source
+     * @param  TType&\UnitEnum  $target
      *
      * @return static
      */
-    public static function make($source, $target): Transition
+    public static function make(\UnitEnum $source, \UnitEnum $target): Transition
     {
         return new static($source, $target);
     }
 
     /**
-     * @param  TType  $source
-     * @param  TType  $target
+     * @param  TType&\UnitEnum  $source
+     * @param  TType&\UnitEnum  $target
      */
-    public function __construct($source, $target)
+    public function __construct(public \UnitEnum $source, public \UnitEnum $target)
     {
-        $this->source = $source;
-        $this->target = $target;
         $this->context = new ContextRepository;
     }
 
@@ -101,12 +85,11 @@ class Transition implements Arrayable, Injectable
     /**
      * Authorize transition using policy ability (or closure).
      *
-     * @param  callable|string|null  $ability
-     * @param  callable|null  $callback
+     * @param  string|callable(Model, Transition): boolean  $ability
      */
-    public function authorizedBy($ability = null, callable $callback = null): self
+    public function authorizedBy(callable|string $ability): static
     {
-        $this->authorization = $ability ?? $callback;
+        $this->authorization = $ability;
 
         return $this;
     }
@@ -114,11 +97,39 @@ class Transition implements Arrayable, Injectable
     /**
      * Hide transition from humans, so only robots can move it.
      */
-    public function hidden(): self
+    public function hidden(): static
     {
         $this->authorization = fn() => false;
 
         return $this;
+    }
+
+    /**
+     * Ability to authorize.
+     *
+     * @return null|string|callable(Model, Transition): boolean
+     */
+    public function authorization(): callable|string|null
+    {
+        return $this->authorization;
+    }
+
+    /**
+     * Check if transition authorized.
+     */
+    public function authorized(): ?static
+    {
+        $allowed = null;
+
+        if ($ability = $this->authorization()) {
+            if (is_string($ability)) {
+                $allowed = Gate::allows($ability, [$this->engine()->model, $this]);
+            } elseif (is_callable($ability)) {
+                $allowed = call_user_func($ability, $this->engine()->model, $this);
+            }
+        }
+
+        return $allowed === false ? null : $this;
     }
 
     /**
@@ -130,9 +141,7 @@ class Transition implements Arrayable, Injectable
     {
         $this->prerequisites()
             ->merge($this->target()->prerequisites())
-            ->each(function ($condition) {
-                call_user_func($condition, $this->engine->model);
-            });
+            ->each(fn(callable $condition) => call_user_func($condition, $this->engine->model));
 
         return $this;
     }
@@ -155,8 +164,8 @@ class Transition implements Arrayable, Injectable
 
         return [
                 'name'   => $this->caption(),
-                'source' => Value::scalar($this->source),
-                'target' => Value::scalar($this->target),
+                'source' => $this->source()->scalar(),
+                'target' => $this->target()->scalar(),
             ]
             + $rules
             + $issues
@@ -171,13 +180,13 @@ class Transition implements Arrayable, Injectable
      */
     public function caption(): string
     {
-        return $this->resolveCaption($this->engine()->model) ?? $this->target()->caption();
+        return $this->resolveCaption($this->engine->model) ?? $this->target()->caption();
     }
 
     public function chronicle(?Model $performer): ?string
     {
         if (is_callable($this->footprint)) {
-            $chronicle = call_user_func($this->footprint, $this->engine()->model, $performer);
+            $chronicle = call_user_func($this->footprint, $this->engine->model, $performer);
         }
 
         return $chronicle ?? $this->target()->chronicle($performer);
@@ -202,12 +211,12 @@ class Transition implements Arrayable, Injectable
     /**
      * Check the transition route.
      *
-     * @param  TType  $source
-     * @param  TType  $target
+     * @param  TType&\UnitEnum  $source
+     * @param  TType&\UnitEnum  $target
      *
      * @return bool
      */
-    public function route($source, $target): bool
+    public function route(\UnitEnum $source, \UnitEnum $target): bool
     {
         return $this->source()->is($source) && $this->target()->is($target);
     }
@@ -215,7 +224,7 @@ class Transition implements Arrayable, Injectable
     /**
      * Transition required to be charged to fire.
      */
-    public function chargeable(Charge $charge): self
+    public function chargeable(Charge $charge): static
     {
         $this->charge = $charge;
 
@@ -231,54 +240,27 @@ class Transition implements Arrayable, Injectable
     }
 
     /**
-     * Ability to authorize.
-     *
-     * @return string|callable|null
-     */
-    public function authorization()
-    {
-        return $this->authorization;
-    }
-
-    /**
-     * Check if transition authorized.
-     */
-    public function authorized(): ?self
-    {
-        $allowed = null;
-
-        if ($ability = $this->authorization()) {
-            if (is_string($ability)) {
-                $allowed = Gate::allows($ability, [$this->engine()->model, $this]);
-            } elseif (is_callable($ability)) {
-                $allowed = call_user_func($ability, $this->engine()->model, $this);
-            }
-        }
-
-        return $allowed === false ? null : $this;
-    }
-
-    /**
      * Get a list of problems with the transition.
      *
-     * @return array<string>
+     * @return array<array-key, string>
      */
     public function issues(): array
     {
         return $this->prerequisites()
             ->merge($this->target()->prerequisites())
-            ->map(function ($condition) {
+            ->filter(function (callable $condition) {
                 try {
                     call_user_func($condition, $this->engine->model);
-                } catch (TransitionFatalException $exception) {
+                    // No problems
+                    return false;
+                } catch (TransitionFatalException) {
                     // Skip
+                    return false;
                 } catch (TransitionRecoverableException $exception) {
                     // Collect only recoverable messages
                     return $exception->getMessage();
                 }
-                return '';
             })
-            ->filter()
             ->values()
             ->toArray();
     }
@@ -288,20 +270,17 @@ class Transition implements Arrayable, Injectable
      *
      * @throws ValidationException
      */
-    public function context(array $context = null): ContextRepository
+    public function withContext(array $context): static
     {
-        if (is_array($context)) {
+        $rules = $this->mergeRules($this->target()->validationRules());
 
-            $rules = $this->mergeRules($this->target()->validationRules());
-
-            if ($rules) {
-                $this->context = new ContextRepository(
-                    validator($context, $rules)->validate()
-                );
-            }
+        if ($rules) {
+            $this->context = new ContextRepository(
+                validator($context, $rules)->validate()
+            );
         }
 
-        return $this->context;
+        return $this;
     }
 
     /**
@@ -310,9 +289,10 @@ class Transition implements Arrayable, Injectable
      * @param  array  $context
      *
      * @return TModel
+     * @throws ValidationException
      */
     public function transit(array $context = [])
     {
-        return $this->engine()->transit($this->target, $context);
+        return $this->engine->transit($this->target, $context);
     }
 }
