@@ -11,7 +11,7 @@ class CreateTransitionHistoryTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('transition_history', function (Blueprint $table) {
             $table->id();
@@ -30,7 +30,7 @@ class CreateTransitionHistoryTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('transition_history');
     }
