@@ -2,12 +2,10 @@
 
 namespace Codewiser\Workflow\Contracts;
 
-use Illuminate\Database\Eloquent\Model;
-
 interface StateEnum
 {
     /**
-     * Human readable state caption.
+     * Human-readable state caption.
      */
     public function caption(): string;
 
