@@ -3,9 +3,9 @@
 namespace Codewiser\Workflow\Exceptions;
 
 /**
- * Indicates, that transition has some problems, that prevents transit.
+ * Indicates that transition has some problems, that prevents transit.
  */
 class TransitionException extends WorkflowException
 {
-
+    //
 }
