@@ -5,11 +5,8 @@ namespace Codewiser\Workflow\Listeners;
 use Codewiser\Workflow\Events\ModelInitialized;
 use Codewiser\Workflow\Events\ModelTransited;
 use Codewiser\Workflow\Models\TransitionHistory;
-use Codewiser\Workflow\State;
 use Codewiser\Workflow\StateMachineEngine;
-use Codewiser\Workflow\Value;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Validation\ValidationException;
 
 class TransitionListener
 {
@@ -28,9 +25,7 @@ class TransitionListener
     {
         $log = $this->newRecordFor($event->model, $event->engine);
 
-        $log->target = Value::scalar(
-            $event->context->target()
-        );
+        $log->target = $event->context->target()->scalar();
 
         $log->context = $event->context->data()->all() ?: null;
 
@@ -41,13 +36,9 @@ class TransitionListener
     {
         $log = $this->newRecordFor($event->model, $event->engine);
 
-        $log->source = Value::scalar(
-            $event->context->source()
-        );
+        $log->source = $event->context->source()->scalar();
 
-        $log->target = Value::scalar(
-            $event->context->target()
-        );
+        $log->target = $event->context->target()->scalar();
 
         $log->context = $event->context->data()->all() ?: null;
 
