@@ -8,13 +8,8 @@ trait HasAttributes
 
     /**
      * Set any additional attribute: color, order, etc.
-     *
-     * @param string $attribute
-     * @param mixed $value
-     *
-     * @return $this
      */
-    public function set(string $attribute, $value): self
+    public function set(string $attribute, mixed $value): static
     {
         $this->additional[$attribute] = $value;
 
@@ -23,6 +18,8 @@ trait HasAttributes
 
     /**
      * Get additional attributes.
+     *
+     * @return array<string, mixed>
      */
     public function additional(): array
     {
