@@ -1,12 +1,10 @@
 <?php
-
 namespace Codewiser\Workflow;
-
 
 /**
  * Workflow blueprint.
  *
- * @template TType
+ * @template TType of \BackedEnum
  */
 abstract class WorkflowBlueprint
 {
@@ -18,16 +16,14 @@ abstract class WorkflowBlueprint
     /**
      * Array of available Model Workflow steps. The first one is initial.
      *
-     * @return array<int,TType|\Codewiser\Workflow\State>
-     * @example [new, review, published, correcting]
+     * @return array<int, TType|\Codewiser\Workflow\State>
      */
     abstract public function states(): array;
 
     /**
      * Array of allowed transitions between states.
      *
-     * @return array<int,array<int,TType|\Codewiser\Workflow\State>|\Codewiser\Workflow\Transition>
-     * @example [[new, review], [review, published], [review, correcting], [correcting, review]]
+     * @return array<int, \Codewiser\Workflow\Transition|array<int, TType|\Codewiser\Workflow\State>>
      */
     abstract public function transitions(): array;
 }
