@@ -5,6 +5,7 @@ namespace Tests;
 use Codewiser\Workflow\Context;
 use Codewiser\Workflow\Events\ModelTransited;
 use Codewiser\Workflow\Example\Article;
+use Codewiser\Workflow\Example\ArticleWithEnum;
 use Codewiser\Workflow\Example\ArticleWorkflow;
 use Codewiser\Workflow\Example\Enum;
 use Codewiser\Workflow\Exceptions\TransitionFatalException;
@@ -24,7 +25,7 @@ class BaseTest extends TestCase
 {
     public function testBasics()
     {
-        $post = new Article();
+        $post = new ArticleWithEnum();
 
         $this->assertNull($post->state, 'State is not initialized');
 
@@ -35,77 +36,41 @@ class BaseTest extends TestCase
 
     public function testStateCollection()
     {
-        $collection = StateCollection::make(['first', 'second']);
+        $collection = StateCollection::make([Enum::new, Enum::review]);
 
         $this->assertCount(2, $collection);
-        $this->assertNotNull($collection->one('first'));
-        $this->assertNotNull($collection->one('second'));
+        $this->assertNotNull($collection->one(Enum::new));
+        $this->assertNotNull($collection->one(Enum::review));
 
         $this->expectException(ItemNotFoundException::class);
-        $collection->one('third');
-    }
-
-    public function testEnumStateCollection()
-    {
-        if (Value::enum_support()) {
-            $collection = StateCollection::make(Enum::cases());
-
-            $this->assertNotNull($collection->one(Enum::new));
-            $this->assertNotNull($collection->one(Enum::review));
-            $this->assertNotNull($collection->one(Enum::published));
-            $this->assertNotNull($collection->one(Enum::correction));
-
-            $this->expectException(ItemNotFoundException::class);
-            $collection->one('third');
-        } else {
-            $this->markTestSkipped('php@8.1 required');
-        }
+        $collection->one(Enum::correction);
     }
 
     public function testTransitionCollection()
     {
-        $collection = TransitionCollection::make([['first', 'second'], ['second', 'first']]);
+        $collection = TransitionCollection::make([[Enum::new, Enum::review], [Enum::review, Enum::published]]);
 
-        $this->assertCount(1, $collection->from('first'));
-        $this->assertCount(1, $collection->from('second'));
-        $this->assertCount(0, $collection->from('third'));
+        $this->assertCount(1, $collection->from(Enum::new));
+        $this->assertCount(1, $collection->from(Enum::review));
+        $this->assertCount(0, $collection->from(Enum::published));
 
-        $this->assertCount(1, $collection->to('first'));
-        $this->assertCount(1, $collection->to('second'));
-        $this->assertCount(0, $collection->to('third'));
-    }
-
-    public function testEnumTransitionCollection()
-    {
-        if (Value::enum_support()) {
-            $collection = TransitionCollection::make([[Enum::new, Enum::review], [Enum::review, Enum::published]]);
-
-            $this->assertCount(1, $collection->from(Enum::new));
-            $this->assertCount(1, $collection->from(Enum::review));
-            $this->assertCount(0, $collection->from(Enum::published));
-            $this->assertCount(0, $collection->from('third'));
-
-            $this->assertCount(0, $collection->to(Enum::new));
-            $this->assertCount(1, $collection->to(Enum::review));
-            $this->assertCount(1, $collection->to(Enum::published));
-            $this->assertCount(0, $collection->to('third'));
-        } else {
-            $this->markTestSkipped('php@8.1 required');
-        }
+        $this->assertCount(1, $collection->to(Enum::review));
+        $this->assertCount(1, $collection->to(Enum::published));
+        $this->assertCount(0, $collection->to(Enum::new));
     }
 
     public function testTransitions()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
         $this->assertCount(2, $post->state()->transitions());
     }
 
     public function testRules()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'review'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::review], true);
 
         $data = $post->state()->transitions()->first()->toArray();
 
@@ -115,8 +80,8 @@ class BaseTest extends TestCase
 
     public function testJson()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
         $data = $post->state()->getTransitionListing()->first()->toArray();
 
@@ -129,31 +94,31 @@ class BaseTest extends TestCase
 
     public function testUniqueTransitions()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
-        $this->assertCount(1, $post->state()->transitions()->to('review'));
-        $this->assertCount(0, $post->state()->transitions()->to('published'));
+        $this->assertCount(1, $post->state()->transitions()->to(Enum::review));
+        $this->assertCount(0, $post->state()->transitions()->to(Enum::published));
     }
 
     public function testRelevantTransitions()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
         $post->state()->transitions()
             ->each(function (Transition $transition) use ($post) {
-                // Assert that every relevant transition starts from current state
+                // Assert that every relevant transition starts from a current state
                 $this->assertEquals($post->state, $transition->source);
             });
     }
 
     public function testTransitRecoverable()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
-        $post->state = 'review';
+        $post->state = Enum::review;
 
         // Observer prevents changing state as the transition has unresolved Recoverable condition
         $this->expectException(TransitionRecoverableException::class);
@@ -162,10 +127,10 @@ class BaseTest extends TestCase
 
     public function testTransitFatal()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'review'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::review], true);
 
-        $post->state = 'published';
+        $post->state = Enum::published;
 
         // Observer prevents changing state as the transition has unresolved Fatal condition
         $this->expectException(TransitionFatalException::class);
@@ -174,20 +139,20 @@ class BaseTest extends TestCase
 
     public function testTransitUnauthorized()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'correction'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::correction], true);
 
         // Transition is not authorized
         $this->expectException(AuthorizationException::class);
-        $post->state()->authorize('review');
+        $post->state()->authorize(Enum::review);
     }
 
     public function testTransitUnknown()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
-        $post->state = 'empty';
+        $post->state = Enum::empty;
 
         // Observer prevents changing state to unknown value
         $this->expectException(ItemNotFoundException::class);
@@ -196,8 +161,8 @@ class BaseTest extends TestCase
 
     public function testToArray()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
         $data = $post->state()->toArray();
 
@@ -213,28 +178,30 @@ class BaseTest extends TestCase
 
     public function testChargeable()
     {
-        $post = new Article();
-        $post->setRawAttributes(['state' => 'new'], true);
+        $post = new ArticleWithEnum();
+        $post->setRawAttributes(['state' => Enum::new], true);
 
         $data = $post->state()->toArray();
         $this->assertArrayHasKey('charge', $data['transitions'][1]);
         $this->assertEquals(0, $data['transitions'][1]['charge']['progress']);
 
-        $post->state()->transit('cumulative');
+        $post->state()->transit(Enum::cumulative);
+        $this->assertFalse($post->state()->is(Enum::cumulative));
         $data = $post->state()->toArray();
         $this->assertEquals(1/3, $data['transitions'][1]['charge']['progress']);
 
-        $post->state()->transit('cumulative');
+        $post->state()->transit(Enum::cumulative);
+        $this->assertFalse($post->state()->is(Enum::cumulative));
         $data = $post->state()->toArray();
         $this->assertEquals(2/3, $data['transitions'][1]['charge']['progress']);
 
-        $post->state()->transit('cumulative');
-        $this->assertTrue($post->state()->is('cumulative'));
+        $post->state()->transit(Enum::cumulative);
+        $this->assertTrue($post->state()->is(Enum::cumulative));
     }
 
     public function testMergeRules()
     {
-        $state = new State('one');
+        $state = new State(Enum::new);
         $state->rules([
             'comment' => 'required|string',
         ]);
